refactor module_parser and overall_parser
