+++ tb_plugin/torch_tb_profiler/profiler/module_parser.py
@@ -2,141 +2,12 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # --------------------------------------------------------------------------
 import sys
-from abc import ABC
 
 from .. import utils
+from .node import OperatorNode
 from .trace import EventTypes
 
 logger = utils.get_logger()
-
-CommunicationOpNameSet = ['nccl:broadcast', 'nccl:all_reduce']
-
-class BaseNode(ABC):
-    def __init__(self, name, start_time, end_time, type, external_id):
-        self.name = name
-        self.start_time = start_time
-        self.end_time = end_time
-        self.type = type
-        self.external_id = external_id  # For consistency check.
-
-    @staticmethod
-    def get_node_argument(event):
-        kwargs = {}
-        kwargs['name'] = event.name
-        kwargs['start_time'] = event.ts
-        kwargs['end_time'] = event.ts + event.duration
-        kwargs['type'] = event.type
-
-        if event.external_id is not None:
-            kwargs['external_id'] = event.external_id
-
-        return kwargs
-
-
-class CommunicationNode(BaseNode):
-    def __init__(self, name, start_time, end_time, type, external_id, input_shape, input_type):
-        super().__init__(name, start_time, end_time, type, external_id)
-        self.input_shape = input_shape
-        self.input_type = input_type
-        self.kernel_ranges = []
-        self.total_time = 0
-        self.real_time = 0
-        self.step_name = None
-
-    @classmethod
-    def create(cls, event, input_shape, input_type):
-        kwargs = BaseNode.get_node_argument(event)
-        return cls(input_shape=input_shape, input_type=input_type, **kwargs)
-
-
-class HostNode(BaseNode):
-    def __init__(self, name, start_time, end_time, type, external_id, device_duration=0):
-        super().__init__(name, start_time, end_time, type, external_id)
-        self.device_duration = device_duration  # Total time of Kernel, GPU Memcpy, GPU Memset. TODO: parallel multi-stream?
-
-
-class OperatorNode(HostNode):
-    # Don't use [] as default parameters
-    # https://stackoverflow.com/questions/1132941/least-astonishment-and-the-mutable-default-argument?page=1&tab=votes#tab-top
-    # https://web.archive.org/web/20200221224620/http://effbot.org/zone/default-values.htm
-    def __init__(self, name, start_time, end_time, type, external_id=None, device_duration=0,
-            children=None, runtimes=None, input_shape=None, input_type=None, call_stack=None, self_host_duration=0, self_device_duration=0):
-        super().__init__(name, start_time, end_time, type, external_id, device_duration)
-        self.children = [] if children is None else children # OperatorNode and ProfilerStepNode.
-        self.runtimes = [] if runtimes is None else runtimes # RuntimeNode
-        self.input_shape = input_shape
-        self.input_type = input_type
-        self.call_stack = call_stack
-        self.self_host_duration = self_host_duration
-        self.self_device_duration = self_device_duration
-
-    def fill_stats(self):
-        # TODO: Replace recursive by using a stack, in case of too deep callstack.
-        for child in self.children:
-            child.fill_stats()
-        for rt in self.runtimes:
-            rt.fill_stats()
-            rt.update_device_op_node(self)
-
-        self.self_host_duration = self.end_time - self.start_time
-        for child in self.children:
-            self.device_duration += child.device_duration
-            self.self_host_duration -= (child.end_time - child.start_time)
-        for rt in self.runtimes:
-            # From PyTorch 1.8 RC1, cpu_self_time does not include runtime's time.
-            # So here we keep consistent with it.
-            self.self_host_duration -= (rt.end_time - rt.start_time)
-            self.device_duration += rt.device_duration
-            self.self_device_duration += rt.device_duration
-
-    def replace_time_by_children(self):
-            self.start_time = next((child.start_time for child in self.children if child.start_time is not None), None)
-            self.end_time = next((child.end_time for child in reversed(self.children) if child.end_time is not None), None)
-
-    @classmethod
-    def create(cls, event, input_shape, input_type, call_stack):
-        kwargs = BaseNode.get_node_argument(event)
-        return cls(input_shape=input_shape, input_type=input_type, call_stack=call_stack, **kwargs)
-
-
-class ProfilerStepNode(OperatorNode):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-
-class RuntimeNode(HostNode):
-    def __init__(self, name, start_time, end_time, type, external_id=None, device_duration=0,
-            device_nodes=None):
-        super().__init__(name, start_time, end_time, type, external_id, device_duration)
-        # One runtime could trigger more than one kernel, such as cudaLaunchCooperativeKernelMultiDevice.
-        self.device_nodes = device_nodes
-
-    def fill_stats(self):
-        if self.device_nodes:
-            for device_node in self.device_nodes:
-                self.device_duration += device_node.end_time - device_node.start_time
-
-    def update_device_op_node(self, node):
-        if self.device_nodes:
-            for device_node in self.device_nodes:
-                device_node.op_node = node
-
-    @classmethod
-    def create(cls, event, device_nodes):
-        kwargs = BaseNode.get_node_argument(event)
-        return cls(device_nodes=device_nodes, **kwargs)
-
-
-class DeviceNode(BaseNode):
-    def __init__(self, name, start_time, end_time, type, external_id=None,
-            op_node=None):
-        super().__init__(name, start_time, end_time, type, external_id)
-        self.op_node = op_node  # The cpu operator that launched it.
-
-    @classmethod
-    def create(cls, event):
-        kwargs = BaseNode.get_node_argument(event)
-        return cls(**kwargs)
 
 
 class OperatorAgg:
@@ -251,49 +122,7 @@ class ModuleParser:
         traverse_node(root_node)
         return root_node
 
-    def parse_events(self, events, context_data):
-
-        def parse_event(event, corrid_to_device, corrid_to_runtime, externalid_to_runtime, tid2list, tid2zero_rt_list):
-            corrid = event.args.get("correlation", None)
-            tid = event.tid
-            if event.type in [EventTypes.KERNEL, EventTypes.MEMCPY, EventTypes.MEMSET]:
-                device_node = DeviceNode.create(event)
-                if corrid in corrid_to_runtime:
-                    rt_node = corrid_to_runtime[corrid]  # Don't pop it because it may be used by next kernel.
-                    if rt_node.device_nodes is None:
-                        rt_node.device_nodes = []
-                    rt_node.device_nodes.append(device_node)
-
-                    # Check the external_id
-                    if rt_node.external_id != device_node.external_id:
-                        logger.warning("Runtime and Device-op have same correlation id but with different external id!")
-                else:
-                    corrid_to_device.setdefault(corrid, []).append(device_node)
-                context_data.device_node_list.append(device_node)
-            elif event.type == EventTypes.RUNTIME:
-                device_nodes = corrid_to_device.pop(corrid, None)
-                rt_node = RuntimeNode.create(event, device_nodes)
-                corrid_to_runtime[corrid] = rt_node
-                externalid_to_runtime.setdefault(rt_node.external_id, []).append(rt_node)
-                # Some runtimes has external_id 0, which will not be correlated to any operator.
-                # So get them and attach them to root node.
-                if rt_node.external_id == 0:
-                    tid2zero_rt_list.setdefault(tid, []).append(rt_node)
-                context_data.runtime_node_list.append(rt_node)
-
-                # check the external_id
-                if device_nodes:
-                    for device_node in device_nodes:
-                        if rt_node.external_id != device_node.external_id:
-                            logger.warning("Runtime and Device-op have same correlation id but with different external id!")
-            elif event.type in [EventTypes.PYTHON, EventTypes.OPERATOR, EventTypes.PROFILER_STEP]:
-                if event.type == EventTypes.PROFILER_STEP:
-                    op_node = ProfilerStepNode.create(event, event.input_shape, event.input_type, None)
-                else:
-                    op_node = OperatorNode.create(event, event.input_shape, event.input_type, event.callstack)
-                if event.name in CommunicationOpNameSet:
-                    context_data.communication_data[op_node.external_id] = CommunicationNode.create(event, op_node.input_shape, op_node.input_type)
-                tid2list.setdefault(tid, []).append(op_node)
+    def aggregate(self, context):
 
         def parse_ops(cpp_op_list):
             def aggregate(key_to_agg, key, op):
@@ -355,34 +184,15 @@ class ModuleParser:
             kernel_list_groupby_name_op = list(name_op_to_agg.values())
             return kernel_list_groupby_name_op
 
-        # For OperatorNode and ProfilerStepNode:
-        #   Use time interval containing relationship to build father-child correlation,
-        #   which is consistent with autograd profiler.
-        # For RuntimeNode:
-        #   Use external_id to build correlation with its father OperatorNode or ProfilerStepNode.
-        #   Because in the case when RuntimeNode has duration 0 and starts at same time as a OperatorNode,
-        #   just use interval containing relationship can't tell it is child or brother of the OperatorNode.
-        tid2list = {}  # value is a list of OperatorNode and ProfilerStepNode. Do not include RuntimeNode
-        tid2zero_rt_list = {}  # value is a list of RuntimeNode with external_id=0. They will be attached to root nodes.
-        corrid_to_device = {}  # value is a list of DeviceNode
-        corrid_to_runtime = {}  # value is a RuntimeNode
-        externalid_to_runtime = {}  # value is a list of RuntimeNode
-        for event in events:
-            parse_event(event, corrid_to_device, corrid_to_runtime, externalid_to_runtime, tid2list, tid2zero_rt_list)
+        tid2list = context.tid2list
+        tid2zero_rt_list = context.tid2zero_rt_list
+        corrid_to_device = context.corrid_to_device
+
         # Kernel that not owned by any operator should also be shown in kernel view
         # when group by "Kernel Name + Op Name".
         for _, device_nodes in corrid_to_device.items():
             self.kernel_list.extend([n for n in device_nodes if n.type == EventTypes.KERNEL])
-        # associate CUDA Runtimes with CPU events
-        for _, op_list in tid2list.items():
-            for op in op_list:
-                runtime_nodes = externalid_to_runtime.pop(op.external_id, [])
-                if runtime_nodes:
-                    op.runtimes.extend(runtime_nodes)
-        for ext_id in externalid_to_runtime:
-            if ext_id != 0:
-                logger.warning("{} Runtime with external id {} don't correlate to any operator!".format(
-                    len(externalid_to_runtime[ext_id]), ext_id))
+
         for tid, op_list in tid2list.items():
             zero_rt_list = tid2zero_rt_list[tid] if tid in tid2zero_rt_list else []
             # Note that when 2 start_time are equal, the one with bigger end_time should be ahead of the other.
