+++ tb_plugin/torch_tb_profiler/profiler/overall_parser.py
@@ -1,32 +1,10 @@
 # -------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # --------------------------------------------------------------------------
-import sys
-from enum import IntEnum
-
 from .. import utils
-from .trace import EventTypes
+from .step_parser import ProfileRole, merge_ranges
 
 logger = utils.get_logger()
-
-
-def merge_ranges(src_ranges, is_sorted=False):
-    merged_ranges = []
-    if len(src_ranges) > 0:
-        if not is_sorted:
-            src_ranges.sort(key=lambda x: x[0])
-        src_id = 0
-        merged_ranges.append(
-            (src_ranges[src_id][0], src_ranges[src_id][1]))
-        for src_id in range(1, len(src_ranges)):
-            dst_id = len(merged_ranges) - 1
-            if src_ranges[src_id][1] > merged_ranges[dst_id][1]:
-                if src_ranges[src_id][0] <= merged_ranges[dst_id][1]:
-                    merged_ranges[dst_id] = (merged_ranges[dst_id][0], src_ranges[src_id][1])
-                else:
-                    merged_ranges.append(
-                        (src_ranges[src_id][0], src_ranges[src_id][1]))
-    return merged_ranges
 
 
 def subtract_ranges_lists(range_list1, range_list2):
@@ -116,9 +94,6 @@ def pop_list(range_list, index):
     return next_item, next_index
 
 
-ProfileRole = IntEnum('ProfileRole', ['Kernel', 'Memcpy', 'Memset', 'Communication', 'Runtime', 'DataLoader', 'CpuOp', 'Other', 'Total'], start=0)
-
-
 class OverallParser(object):
     class Costs:
         def __init__(self):
@@ -174,165 +149,41 @@ class OverallParser(object):
             self.other = 0
 
     def __init__(self):
-        # we could not use [[]] * len here since they all point to same memory
-        # https://stackoverflow.com/questions/12791501/python-initializing-a-list-of-lists
-        # https://stackoverflow.com/questions/240178/list-of-lists-changes-reflected-across-sublists-unexpectedly
-        self.role_ranges = [[] for _ in range(ProfileRole.Total - 1)]
-        self.steps = []
-        self.steps_names = []
-
-        self.min_ts = sys.maxsize
-        self.max_ts = -sys.maxsize - 1
         self.steps_costs = []
         self.avg_costs = OverallParser.Costs()
         self.communication_overlap = []
-        self.comm_node_list = []
 
-    # Update self.steps considering device side events launched by each host side step.
-    # Update self.steps_names if some tail steps are removed.
-    def update_steps_consider_device_side(self, runtime_node_list, device_node_list):
-        runtime_node_list = sorted(runtime_node_list, key=lambda x: x.start_time)
-        # Make sure self.steps is sorted by time.
-        self.steps = sorted(self.steps, key=lambda x: x[0])
-        # Use similar code with two-way merge to get all runtimes inside each host-side step span,
-        # then record each step's min kernel start time and max kernel end time:
-        steps_device = [(sys.maxsize, -sys.maxsize - 1)] * len(self.steps)
-        steps_matched_device_nodes = [0] * len(self.steps)
-        i_step = 0
-        i_runtime = 0
-        step_device_min_ts = sys.maxsize
-        step_device_max_ts = -sys.maxsize - 1
-        matched_device_nodes = set()
-        while i_step < len(self.steps) and i_runtime < len(runtime_node_list):
-            step_host_start_time = self.steps[i_step][0]
-            step_host_end_time = self.steps[i_step][1]
-            if runtime_node_list[i_runtime].start_time < step_host_start_time:
-                # This runtime is ahead of or intersects with this step span. Skip this runtime.
-                i_runtime += 1
-            elif runtime_node_list[i_runtime].end_time <= step_host_end_time:
-                # and runtime_node_list[i_runtime].start_time >= step_host_start_time
-                # This runtime is inside this step span. Scan its device_nodes.
-                rt = runtime_node_list[i_runtime]
-                if rt.device_nodes is not None:
-                    for device_node in rt.device_nodes:
-                        step_device_min_ts = min(device_node.start_time, step_device_min_ts)
-                        step_device_max_ts = max(device_node.end_time, step_device_max_ts)
-                        matched_device_nodes.add(device_node)
-                        steps_matched_device_nodes[i_step] += 1
-                i_runtime += 1
-            elif runtime_node_list[i_runtime].start_time < step_host_end_time:
-                # and runtime_node_list[i_runtime].end_time > step_host_end_time
-                # This runtime intersects with this step span. Skip this runtime.
-                i_runtime += 1
-            else:
-                # runtime_node_list[i_runtime].start_time >= step_host_end_time
-                # This runtime starts after this step's end. Record and move forward this step.
-                steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
-                i_step += 1
-                step_device_min_ts = sys.maxsize
-                step_device_max_ts = -sys.maxsize - 1
-        while i_step < len(self.steps):
-            # This step doesn't launch any device side event, just assign it as empty.
-            steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
-            step_device_min_ts = sys.maxsize
-            step_device_max_ts = -sys.maxsize - 1
-            i_step += 1
-        # Change step time to device side on the condition that any step have device time.
-        is_use_gpu = (len(matched_device_nodes) > 0)
-        if is_use_gpu:
-            prev_step_end_time = self.steps[0][0]
-            if steps_device[0][0] != sys.maxsize:  # When step 0 has device event.
-                for device_node in device_node_list:
-                    if device_node not in matched_device_nodes:
-                        # Now this device_node is not launched inside any step span.
-                        if device_node.end_time < steps_device[0][0]:
-                            prev_step_end_time = max(prev_step_end_time, device_node.end_time)
-            for i_step in range(len(self.steps)):
-                step_start_time = max(prev_step_end_time, self.steps[i_step][0])
-                step_end_time = self.steps[i_step][1]
-                if steps_device[i_step][0] == sys.maxsize:  # When step i_step has no device event.
-                    # Assign to step_start_time when kernel is behind host step end.
-                    step_end_time = max(step_end_time, step_start_time)
-                else:
-                    step_end_time = max(step_end_time, steps_device[i_step][1])
-                    if step_end_time < step_start_time:
-                        logger.warning(
-                            "Abnormal step_end_time of step {}: [{}, {}]".format(
-                                i_step, step_start_time, step_end_time))
-                        step_end_time = step_start_time
-                self.steps[i_step] = (step_start_time, step_end_time)  # Update step time considering device side.
-                prev_step_end_time = step_end_time
-
-        is_remove_tail_steps = True  # TODO: Use tensorboard argument instead.
-        if is_use_gpu and len(self.steps) > 1 and is_remove_tail_steps:
-            i_step = len(self.steps) - 1
-            while i_step >= 0:
-                if steps_matched_device_nodes[i_step] > 0:
-                    break
-                i_step -= 1
-            if i_step >= 0:
-                keep_steps = i_step + 1
-                if i_step > 0 and steps_matched_device_nodes[i_step - 1] * 0.8 > steps_matched_device_nodes[i_step]:
-                    keep_steps = i_step
-                if keep_steps < len(self.steps):
-                    logger.warning(
-                        "Remove the last {} steps from overview. "
-                        "Because the profiler may fail to capture all the kernels launched by these steps.".format(
-                            len(self.steps) - keep_steps
-                        ))
-                    self.steps = self.steps[:keep_steps]
-                    self.steps_names = self.steps_names[:keep_steps]
-
-
-    def parse_events(self, events, context_data):
-        logger.debug("Overall, parse events")
-        for event in events:
-            self.parse_event(event, context_data.communication_data)
-
-        if len(self.steps) == 0:
-            self.steps.append((self.min_ts, self.max_ts))
-            self.steps_names.append("0")
-        self.update_steps_consider_device_side(context_data.runtime_node_list, context_data.device_node_list)
-
-        for i in range(len(self.role_ranges)):
-            self.role_ranges[i] = merge_ranges(self.role_ranges[i])
-
+    def aggregate(self, node_parser, step_parser):
         logger.debug("Overall, statistics")
-        global_stats = OverallParser.Statistics.create_statistics(self.steps, self.role_ranges)
-        comm_kernel_overlap = intersection_ranges_lists(self.role_ranges[ProfileRole.Kernel], self.role_ranges[ProfileRole.Communication])
+        global_stats = OverallParser.Statistics.create_statistics(step_parser.steps, step_parser.role_ranges)
+        comm_kernel_overlap = intersection_ranges_lists(step_parser.role_ranges[ProfileRole.Kernel], step_parser.role_ranges[ProfileRole.Communication])
 
         logger.debug("Overall, aggregation")
-        valid_steps = len(self.steps)
+        valid_steps = len(step_parser.steps)
         for i in range(valid_steps):
-            steps_stat = global_stats.intersection_with_step(self.steps[i])
-            self.steps_costs.append(OverallParser.Costs.calculate_costs(steps_stat, self.steps[i]))
+            steps_stat = global_stats.intersection_with_step(step_parser.steps[i])
+            self.steps_costs.append(OverallParser.Costs.calculate_costs(steps_stat, step_parser.steps[i]))
             for cost_index in range(len(self.avg_costs.costs)):
                 self.avg_costs.costs[cost_index] += self.steps_costs[i].costs[cost_index]
 
             comm_costs = OverallParser.StepCommunicationCosts()
-            comm_costs.overlap = get_ranges_sum(intersection_ranges_lists([self.steps[i]], comm_kernel_overlap))
-            comm_costs.computation = get_ranges_sum(intersection_ranges_lists([self.steps[i]], self.role_ranges[ProfileRole.Kernel]))
-            comm_costs.communication = get_ranges_sum(intersection_ranges_lists([self.steps[i]], self.role_ranges[ProfileRole.Communication]))
+            comm_costs.overlap = get_ranges_sum(intersection_ranges_lists([step_parser.steps[i]], comm_kernel_overlap))
+            comm_costs.computation = get_ranges_sum(intersection_ranges_lists([step_parser.steps[i]], step_parser.role_ranges[ProfileRole.Kernel]))
+            comm_costs.communication = get_ranges_sum(intersection_ranges_lists([step_parser.steps[i]], step_parser.role_ranges[ProfileRole.Communication]))
             comm_costs.other = self.steps_costs[i].costs[ProfileRole.Total] + comm_costs.overlap - comm_costs.computation - comm_costs.communication
             self.communication_overlap.append(comm_costs)
 
         for i in range(len(self.avg_costs.costs)):
             self.avg_costs.costs[i] /= valid_steps
 
-        # Sort the communication node according the start time, this is for correlating communication node between workers
-        for comm_node in context_data.communication_data.values():
-            comm_node.kernel_ranges.sort(key=lambda x: (x[0], -x[1]))
-            self.comm_node_list.append(comm_node)
-        self.comm_node_list.sort(key=lambda x: (x.start_time, -x.end_time))
-
         # Find each communication node belong to which step
         index = 0
-        for comm_node in self.comm_node_list:
+        for comm_node in node_parser.comm_node_list:
             while index < valid_steps:
-                if comm_node.start_time >= self.steps[index][0] and comm_node.end_time <= self.steps[index][1]:
-                    comm_node.step_name = self.steps_names[index]
+                if comm_node.start_time >= step_parser.steps[index][0] and comm_node.end_time <= step_parser.steps[index][1]:
+                    comm_node.step_name = step_parser.steps_names[index]
                     break
-                elif comm_node.start_time >= self.steps[index][1]:
+                elif comm_node.start_time >= step_parser.steps[index][1]:
                     index += 1
                 else:
                     logger.error("Found a communication op not belong to any step.")
@@ -340,37 +191,3 @@ class OverallParser(object):
             if index >= valid_steps:
                 logger.error("Found communication ops not belong to any step. ")
                 break
-
-    def parse_event(self, event, communication_data):
-        ts = event.ts
-        dur = event.duration
-        evt_type = event.type
-        external_id = event.external_id
-        if evt_type == EventTypes.KERNEL:
-            if external_id in communication_data:
-                self.role_ranges[ProfileRole.Communication].append((ts, ts + dur))
-                communication_data[external_id].kernel_ranges.append((ts, ts + dur))
-                communication_data[external_id].total_time += dur
-            else:
-                self.role_ranges[ProfileRole.Kernel].append((ts, ts + dur))
-        elif evt_type == EventTypes.MEMCPY:
-            self.role_ranges[ProfileRole.Memcpy].append((ts, ts + dur))
-        elif evt_type == EventTypes.MEMSET:
-            self.role_ranges[ProfileRole.Memset].append((ts, ts + dur))
-        elif evt_type == EventTypes.RUNTIME:
-            self.role_ranges[ProfileRole.Runtime].append((ts, ts + dur))
-        elif evt_type == EventTypes.OPERATOR and event.name.startswith("enumerate(DataLoader)#") \
-                and event.name.endswith(".__next__"):
-            self.role_ranges[ProfileRole.DataLoader].append((ts, ts + dur))
-        elif event.type == EventTypes.PROFILER_STEP:
-            self.steps.append((ts, ts + dur))
-            self.steps_names.append(str(event.step))
-        elif evt_type in [EventTypes.PYTHON, EventTypes.OPERATOR]:
-            self.role_ranges[ProfileRole.CpuOp].append((ts, ts + dur))
-
-        # Record host side min and max time.
-        if evt_type in [EventTypes.PYTHON, EventTypes.OPERATOR, EventTypes.PROFILER_STEP]:
-            if ts < self.min_ts:
-                self.min_ts = ts
-            if ts + dur > self.max_ts:
-                self.max_ts = ts + dur
